# YJIT: Add compilation time counter