This PR adds a counter of cumulative time taken for compiling JIT code. Similar to `GC.stat(:time)`, if you call this before and after your request, you can measure how long you spent in YJIT compilation for each request.

When we compare multiple Ruby versions, we could monitor `GC.stat(:time)` and this to check how much impact each of GC and YJIT had on their difference in performance.

### Example

| benchmark | compile_time_ms |
|:------------|:------------------|
| 30k_ifelse | 1,234ms |
| 30k_methods | 429ms |
| railsbench | 621ms |
| lobsters | 1,699ms |