@k0kubun 2023-09-12
> could you also measure if there is any impact on total run time for the first iteration of e.g. 30k_ifelse

I ran one iteration of the following benchmarks:

### --yjit

```
before: ruby 3.3.0dev (2023-09-11T17:09:22Z master 1ae5dd6f32) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-12T16:02:16Z yjit-compile-time 42b9e18d96) +YJIT [x86_64-linux]

-----------  -----------  ----------  ----------  ----------  -------------  ------------
bench        before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
lobsters     1738.2       0.0         1722.0      0.0         1.01           1.01
railsbench   1498.7       0.0         1493.5      0.0         1.00           1.00
30k_ifelse   1345.2       0.0         1327.1      0.0         1.01           1.01
-----------  -----------  ----------  ----------  ----------  -------------  ------------
```

### --yjit-stats

```
before: ruby 3.3.0dev (2023-09-11T17:09:22Z master 1ae5dd6f32) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-12T16:02:16Z yjit-compile-time 42b9e18d96) +YJIT [x86_64-linux]

----------  -----------  ----------  ----------  ----------  -------------  ------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
lobsters    1918.4       0.0         1909.9      0.0         1.00           1.00
railsbench  1606.6       0.0         1603.7      0.0         1.00           1.00
30k_ifelse  1725.2       0.0         1705.3      0.0         1.01           1.01
----------  -----------  ----------  ----------  ----------  -------------  ------------
```

This PR doesn't seem to impact the 1st itr performance whether this feature is enabled or not.